Add files via upload
Domain: Industrial safety. NLP based Chatbot
Context: The database comes from one of the biggest industries in Brazil and in the world. It is an urgent need for industries/companies around the globe to understand why employees still suffer some injuries/accidents in plants. Sometimes they also die in such an environment.
Data Description: This database basically records accidents from 12 different plants in 03 different countries which every line in the data is an occurrence of an accident. Columns description: ‣ Data: timestamp or time/date information ‣ 
Countries: which country the accident occurred (anonymised) ‣ Local: the city where the manufacturing plant is located (anonymised) 
Industry sector: which sector the plant belongs to
Accident level: from I to VI, it registers how severe was the accident (I means not severe but VI means very severe)
Potential Accident Level: Depending on the Accident Level, the database also registers how severe the accident could have been (due to other factors involved in the accident)
Gender: if the person is male of female ‣ Employee or Third Party: if the injured person is an employee or a third party 
Critical Risk: some description of the risk involved in the accident Description: Detailed description of how the accident happened
Description: Detailed description of how the accident happened